+++ ceammc/ext/src/preset/preset_storage.h
@@ -20,6 +20,7 @@ public:
     void m_write(t_symbol*, const AtomList& fname);
     void m_read(t_symbol*, const AtomList& fname);
     void m_update(t_symbol*, const AtomList&);
+    void m_duplicate(t_symbol*, const AtomList& l);
 
     std::string makeDefaultPresetPath() const;
 };